+++ src/main/java/com/xml/fixer/Comment.java
@@ -2,21 +2,21 @@ package com.xml.fixer;
 
 public final class Comment {
 
-    private final long start;
-    private final long end;
+    private final int start;
+    private final int end;
     private final String text;
 
-    public Comment(final long start, final long end, final String text) {
+    public Comment(final int start, final int end, final String text) {
         this.start = start;
         this.end = end;
         this.text = text;
     }
 
-    public long getStart() {
+    public int getStart() {
         return start;
     }
 
-    public long getEnd() {
+    public int getEnd() {
         return end;
     }
